requirements.txt
Flask==3.1.3
gunicorn==25.1.0

requests==2.32.5
structlog==25.1.0
sentencepiece==0.2.0
torch
transformers
google-cloud-storage==3.2.0

google-auth==2.49.0
